fix: 🚑️ Fix merge errors
Fix import cycle error
